neurio.go - Add comment about the apparent issue with the watt-second values reporting 2x
The issue seems likely due to a sampling issue. The time reported for each sample has
a granularity of seconds but perhaps two samples are taken, one towards the start of
a given second and the other towards the end so the watt-second delta is nearly two
seconds but if rounding down when reporting the time it appears that the samples are
one second apart.